consolidate misc package
