--- pydl/pydlutils/misc/__init__.py
@@ -1,11 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the misc directory of idlutils.
-"""
-from .decode_mixed import decode_mixed
-from .djs_laxisgen import djs_laxisgen
-from .djs_laxisnum import djs_laxisnum
-# from .djs_readcol import djs_readcol
-from .hogg_iau_name import hogg_iau_name
-from .struct_print import struct_print
